corrected annoying warnings produced by msvc
svn path=/trunk/iCub/; revision=6704

+++ src/modules/vergence/include/iCub/disparityTool.h
@@ -79,7 +79,7 @@ public:
         if (index < 0)
             index = 0;
         else if (index > (max-min))
-            index = max-min;			  
+            index = (int)(max-min);
         return _shiftFunctionInv[index]; 
     }
